MAGETWO-56126: Login failed after new custom attribute was added
- handling unlock with a specialized model that can save only auth data for the customer
